Added Logic for Events to handle DB writes
The issue I'm looking to solve is that currently
if I simply loop and write the data it requires builing
and throwing away a writer object so many times it simply
cannot keep up with how fast we want to write the data. So,
my first idea is to simply use an event handled to write the
changes if/when necessary. I've implemeted a new DBWriterArgs
class that will hold the required args to be passed to the
event to write to the database.