Add if statements for _ret check
While running the information gatherer in the background
I ran into an error where a call through an out of
bounds exception while looping through an array
because the short variable returned by reference
was not changed which caused it to be short.MaxValue.
Implementing the check for _ret == Focas1.EW_OK should resolve
that issue at a high level. If the call fails it should
only be due to a connection failure and so the next update
should implement that fix.